+++ src/main/java/net/brdle/collectorsreap/common/item/food/Nutrition.java
@@ -295,6 +295,11 @@ public class Nutrition {
 		.nutrition(6).saturationMod(0.4F)
 		.effect(() -> new MobEffectInstance(CREffects.REBOUND.get(), 1600, 2), 1F)
 		.build();
+	public static final FoodProperties TRICOLOR_ICE_CREAM = (new FoodProperties.Builder())
+		.nutrition(6).saturationMod(0.4F)
+		.effect(() -> new MobEffectInstance(CREffects.REBOUND.get(), 1600, 1), 1F)
+		.effect(() -> new MobEffectInstance(ModCompat.getVanillaScent(), 200, 0), 1F)
+		.build();
 	public static final FoodProperties LIME_MILKSHAKE = (new FoodProperties.Builder())
 		.nutrition(2).saturationMod(1.5F).alwaysEat()
 		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 1), 1F)